Changed all constraint configuration into SnapKit - Added PodFile to configure SnapKit (Run command "pod install" di directory projectnya, trus nanti buka projectnya dari yang .xcworkspace) - Rapihin foldering dikit
