+++ ViewControllers/SheetPresentationController/SheetPresentationController.xib
@@ -20,15 +20,14 @@
             <rect key="frame" x="0.0" y="0.0" width="393" height="852"/>
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
             <subviews>
-                <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" style="plain" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="-1" estimatedSectionHeaderHeight="-1" sectionFooterHeight="-1" estimatedSectionFooterHeight="-1" translatesAutoresizingMaskIntoConstraints="NO" id="7mP-4y-m9L">
+                <tableView clipsSubviews="YES" contentMode="scaleToFill" fixedFrame="YES" alwaysBounceVertical="YES" style="plain" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="-1" estimatedSectionHeaderHeight="-1" sectionFooterHeight="-1" estimatedSectionFooterHeight="-1" translatesAutoresizingMaskIntoConstraints="NO" id="7mP-4y-m9L">
                     <rect key="frame" x="16" y="135" width="361" height="704"/>
+                    <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
                     <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-                    <constraints>
-                        <constraint firstAttribute="height" constant="704" id="tHN-2F-SiZ"/>
-                    </constraints>
                 </tableView>
-                <searchBar contentMode="redraw" placeholder="Enter your destination" translatesAutoresizingMaskIntoConstraints="NO" id="7xZ-up-FkU">
+                <searchBar contentMode="redraw" fixedFrame="YES" placeholder="Enter your destination" translatesAutoresizingMaskIntoConstraints="NO" id="7xZ-up-FkU">
                     <rect key="frame" x="16" y="59" width="361" height="56"/>
+                    <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
                     <textInputTraits key="textInputTraits"/>
                     <connections>
                         <outlet property="delegate" destination="-1" id="tby-X5-SUa"/>
@@ -37,14 +36,6 @@
             </subviews>
             <viewLayoutGuide key="safeArea" id="fnl-2z-Ty3"/>
             <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-            <constraints>
-                <constraint firstAttribute="trailing" secondItem="7xZ-up-FkU" secondAttribute="trailing" constant="16" id="62I-ar-cWc"/>
-                <constraint firstItem="7mP-4y-m9L" firstAttribute="top" secondItem="7xZ-up-FkU" secondAttribute="bottom" constant="20" id="LS3-jf-Fj8"/>
-                <constraint firstItem="7xZ-up-FkU" firstAttribute="leading" secondItem="i5M-Pr-FkT" secondAttribute="leading" constant="16" id="Ps5-p6-KrZ"/>
-                <constraint firstAttribute="trailing" secondItem="7mP-4y-m9L" secondAttribute="trailing" constant="16" id="aEg-GG-IRx"/>
-                <constraint firstItem="7xZ-up-FkU" firstAttribute="top" secondItem="fnl-2z-Ty3" secondAttribute="top" id="qEG-xw-lJv"/>
-                <constraint firstItem="7mP-4y-m9L" firstAttribute="leading" secondItem="i5M-Pr-FkT" secondAttribute="leading" constant="16" id="rT8-rJ-x8M"/>
-            </constraints>
             <point key="canvasLocation" x="107.63358778625954" y="19.718309859154932"/>
         </view>
     </objects>
